add pollIntervalMultiplier
with no multiplication by default (ms) it will break all current apps
using a setting for pollInterval value(s) (as these are most (if not
all) specified in seconds, not milliseconds, to keep it user friendly)

this will add a multiplication option to convert the value.
can't add the *1000 multiplication standard, as that breaks current
SDKv2 apps (e.g. Samsung Z-Wave Locks)
and this gives the developer more posibilities, might he/she want to
have a minutes/hours convertion