Adds ZigBee system capability 'windowcoverings_set'
The system capability `windowcoverings_set` uses the `closuresWindowCovering` cluster. It uses the `goToLiftPercentage` command and reads and binds for reporting on the `currentPositionLiftPercentage` attribute. Additionally, the `setParser` method exposes a `_reportDebounceEnabled` instance variable which can be used by the device driver to determine if reports send by the device are generated because of a set command by Homey or because of external inputs.